# storage/api-docs/api-docs.json
{"openapi":"3.0.0","info":{"title":"Restful API Laravel","version":"1.1.2"},"paths":{"\/api\/register":{"post":{"tags":["User"],"summary":"Register","operationId":"register","parameters":[{"name":"company","in":"query","required":true,"schema":{"type":"string"}},{"name":"represent","in":"query","required":true,"schema":{"type":"string"}},{"name":"email","in":"query","required":false,"schema":{"type":"string"}},{"name":"tel","in":"query","required":true,"schema":{"type":"string"}},{"name":"address","in":"query","required":true,"schema":{"type":"string"}},{"name":"address_number","in":"query","required":true,"schema":{"type":"string"}},{"name":"postcode","in":"query","required":true,"schema":{"type":"string"}},{"name":"certificate_image","in":"query","required":false,"schema":{"type":"file"}},{"name":"services","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"Success","content":{"application\/json":{}}}}}},"\/api\/login":{"post":{"tags":["User"],"summary":"Login","operationId":"login","parameters":[{"name":"user_name","in":"query","required":true,"schema":{"type":"string"}},{"name":"password","in":"query","required":true,"schema":{"type":"string","format":"password"}}],"responses":{"200":{"description":"Success","content":{"application\/json":{}}}}}},"\/api\/profile":{"post":{"tags":["User"],"summary":"Profile","operationId":"profile","parameters":[{"name":"company","in":"query","required":true,"schema":{"type":"string"}},{"name":"represent","in":"query","required":true,"schema":{"type":"string"}},{"name":"email","in":"query","required":false,"schema":{"type":"string"}},{"name":"tel","in":"query","required":true,"schema":{"type":"string"}},{"name":"address","in":"query","required":true,"schema":{"type":"string"}},{"name":"address_number","in":"query","required":true,"schema":{"type":"string"}},{"name":"postcode","in":"query","required":true,"schema":{"type":"string"}},{"name":"certificate_image","in":"query","required":false,"schema":{"type":"file"}},{"name":"user_funds","in":"query","required":false,"schema":{"type":"string"}},{"name":"number_of_staff","in":"query","required":false,"schema":{"type":"string"}},{"name":"business_type","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"Success","content":{"application\/json":{}}}}}}},"components":{"securitySchemes":{"jwt":{"type":"apiKey","description":"JWT security for API","name":"access-token","in":"header"}}}}